Do not decode stack frames on error value construction - instead delay the decoding until the symbols are required
