+++ trace.go
@@ -19,10 +19,7 @@ limitations under the License.
 package trace
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"html/template"
 	"strings"
 	"sync/atomic"
 
@@ -73,34 +70,12 @@ func Unwrap(err error) error {
 	return err
 }
 
-// UserMessager returns a user message associated with the error
-type UserMessager interface {
-	// UserMessage returns the user message associated with the error if any
-	UserMessage() string
-}
+var UserMessage = internal.UserMessage
+var DebugReport = internal.DebugReport
 
-// ErrorWrapper wraps another error
-type ErrorWrapper interface {
-	// OrigError returns the wrapped error
-	OrigError() error
-}
-
-// DebugReporter formats an error for display
-type DebugReporter interface {
-	// DebugReport formats an error for display
-	DebugReport() string
-}
-
-// UserMessage returns user-friendly part of the error
-func UserMessage(err error) string {
-	if err == nil {
-		return ""
-	}
-	if wrap, ok := err.(UserMessager); ok {
-		return wrap.UserMessage()
-	}
-	return err.Error()
-}
+type UserMessager = internal.UserMessager
+type ErrorWrapper = internal.ErrorWrapper
+type DebugReporter = internal.DebugReporter
 
 // UserMessageWithFields returns user-friendly error with key-pairs as part of the message
 func UserMessageWithFields(err error) string {
@@ -117,18 +92,6 @@ func UserMessageWithFields(err error) string {
 			kvps = append(kvps, fmt.Sprintf("%v=%q", k, v))
 		}
 		return fmt.Sprintf("%v %v", strings.Join(kvps, " "), wrap.UserMessage())
-	}
-	return err.Error()
-}
-
-// DebugReport returns debug report with all known information
-// about the error including stack trace if it was captured
-func DebugReport(err error) string {
-	if err == nil {
-		return ""
-	}
-	if reporter, ok := err.(DebugReporter); ok {
-		return reporter.DebugReport()
 	}
 	return err.Error()
 }
@@ -169,180 +132,20 @@ func Errorf(format string, args ...interface{}) (err error) {
 func Fatalf(format string, args ...interface{}) error {
 	if IsDebug() {
 		panic(fmt.Sprintf(format, args...))
-	} else {
-		return Errorf(format, args...)
 	}
+	return Errorf(format, args...)
 }
 
 func newTrace(err error, depth int) *TraceErr {
-	traces := internal.CaptureTraces(depth)
-	return &TraceErr{Err: err, Traces: traces}
+	return internal.Wrap(err, depth+1)
 }
 
 type Traces = internal.Traces
 
-type Trace = internal.Trace
-
-// MarshalJSON marshals this error as JSON-encoded payload
-func (r *TraceErr) MarshalJSON() ([]byte, error) {
-	if r == nil {
-		return nil, nil
-	}
-	type marshalableError TraceErr
-	err := marshalableError(*r)
-	err.Err = &RawTrace{Message: r.Err.Error()}
-	return json.Marshal(err)
-}
-
-// TraceErr contains error message and some additional
-// information about the error origin
-type TraceErr struct {
-	// Err is the underlying error that TraceErr wraps
-	Err error `json:"error"`
-	// Traces is a slice of stack trace entries for the error
-	Traces `json:"traces,omitempty"`
-	// Message is an optional message that can be wrapped with the original error.
-	//
-	// This field is obsolete, replaced by messages list below.
-	Message string `json:"message,omitempty"`
-	// Messages is a list of user messages added to this error.
-	Messages []string `json:"messages,omitempty"`
-	// Fields is a list of key-value-pairs that can be wrapped with the error to give additional context
-	Fields map[string]interface{} `json:"fields,omitempty"`
-}
+type TraceErr = internal.TraceErr
 
 // Fields maps arbitrary keys to values inside an error
 type Fields map[string]interface{}
-
-// Error returns the error message this trace describes.
-// Implements error
-func (r *RawTrace) Error() string {
-	return r.Message
-}
-
-// RawTrace describes the error trace on the wire
-type RawTrace struct {
-	// Err specifies the original error
-	Err json.RawMessage `json:"error,omitempty"`
-	// Traces lists the stack traces at the moment the error was recorded
-	Traces `json:"traces,omitempty"`
-	// Message specifies the optional user-facing message
-	Message string `json:"message,omitempty"`
-	// Messages is a list of user messages added to this error.
-	Messages []string `json:"messages,omitempty"`
-	// Fields is a list of key-value-pairs that can be wrapped with the error to give additional context
-	Fields map[string]interface{} `json:"fields,omitempty"`
-}
-
-// AddUserMessage adds user-friendly message describing the error nature
-func (e *TraceErr) AddUserMessage(formatArg interface{}, rest ...interface{}) *TraceErr {
-	newMessage := fmt.Sprintf(fmt.Sprintf("%v", formatArg), rest...)
-	e.Messages = append(e.Messages, newMessage)
-	return e
-}
-
-// AddFields adds the given map of fields to the error being reported
-func (e *TraceErr) AddFields(fields map[string]interface{}) *TraceErr {
-	if e.Fields == nil {
-		e.Fields = make(map[string]interface{}, len(fields))
-	}
-	for k, v := range fields {
-		e.Fields[k] = v
-	}
-	return e
-}
-
-// AddField adds a single field to the error wrapper as context for the error
-func (e *TraceErr) AddField(k string, v interface{}) *TraceErr {
-	if e.Fields == nil {
-		e.Fields = make(map[string]interface{}, 1)
-	}
-	e.Fields[k] = v
-	return e
-}
-
-// UserMessage returns user-friendly error message
-func (e *TraceErr) UserMessage() string {
-	if len(e.Messages) > 0 {
-		// Format all collected messages in the reverse order, with each error
-		// on its own line with appropriate indentation so they form a tree and
-		// it's easy to see the cause and effect.
-		var buf bytes.Buffer
-		fmt.Fprintln(&buf, e.Messages[len(e.Messages)-1])
-		index, indent := len(e.Messages)-1, 1
-		for ; index > 0; index, indent = index-1, indent+1 {
-			fmt.Fprintf(&buf, "%v%v\n", strings.Repeat("\t", indent), e.Messages[index-1])
-		}
-		fmt.Fprintf(&buf, "%v%v", strings.Repeat("\t", indent), UserMessage(e.Err))
-		return buf.String()
-	}
-	if e.Message != "" {
-		// For backwards compatibility return the old user message if it's present.
-		return e.Message
-	}
-	return UserMessage(e.Err)
-}
-
-// DebugReport returns developer-friendly error report
-func (e *TraceErr) DebugReport() string {
-	var buf bytes.Buffer
-	err := reportTemplate.Execute(&buf, errorReport{
-		OrigErrType:    fmt.Sprintf("%T", e.Err),
-		OrigErrMessage: e.Err.Error(),
-		Fields:         e.Fields,
-		StackTrace:     e.Traces.String(),
-		UserMessage:    e.UserMessage(),
-	})
-	if err != nil {
-		return fmt.Sprint("error generating debug report: ", err.Error())
-	}
-	return buf.String()
-}
-
-// Error returns user-friendly error message when not in debug mode
-func (e *TraceErr) Error() string {
-	return e.UserMessage()
-}
-
-func (e *TraceErr) GetFields() map[string]interface{} {
-	return e.Fields
-}
-
-// Unwrap returns the error this TraceErr wraps. The returned error may also
-// wrap another one, Unwrap doesn't recursively get the inner-most error like
-// OrigError does.
-func (e *TraceErr) Unwrap() error {
-	return e.Err
-}
-
-// OrigError returns original wrapped error
-func (e *TraceErr) OrigError() error {
-	err := e.Err
-	// this is not an endless loop because I'm being
-	// paranoid, this is a safe protection against endless
-	// loops
-	for i := 0; i < maxHops; i++ {
-		newerr, ok := err.(Error)
-		if !ok {
-			break
-		}
-		next := newerr.OrigError()
-		if next == nil || next == err {
-			break
-		}
-		err = next
-	}
-	return err
-}
-
-// GoString formats this trace object for use with
-// with the "%#v" format string
-func (e *TraceErr) GoString() string {
-	return e.DebugReport()
-}
-
-// maxHops is a max supported nested depth for errors
-const maxHops = 50
 
 // Error is an interface that helps to adapt usage of trace in the code
 // When applications define new error types, they can implement the interface
@@ -444,75 +247,3 @@ func IsAggregate(err error) bool {
 	_, ok := Unwrap(err).(Aggregate)
 	return ok
 }
-
-// wrapProxy wraps the specified error as a new error trace
-func wrapProxy(err error) Error {
-	if err == nil {
-		return nil
-	}
-	return proxyError{
-		// Do not include ReadError in the trace
-		TraceErr: newTrace(err, 3),
-	}
-}
-
-// DebugReport formats the underlying error for display
-// Implements DebugReporter
-func (r proxyError) DebugReport() string {
-	var wrappedErr *TraceErr
-	var ok bool
-	if wrappedErr, ok = r.TraceErr.Err.(*TraceErr); !ok {
-		return DebugReport(r.TraceErr)
-	}
-	var buf bytes.Buffer
-	//nolint:errcheck
-	reportTemplate.Execute(&buf, errorReport{
-		OrigErrType:    fmt.Sprintf("%T", wrappedErr.Err),
-		OrigErrMessage: wrappedErr.Err.Error(),
-		Fields:         wrappedErr.Fields,
-		StackTrace:     wrappedErr.Traces.String(),
-		UserMessage:    wrappedErr.UserMessage(),
-		Caught:         r.TraceErr.Traces.String(),
-	})
-	return buf.String()
-}
-
-// GoString formats this trace object for use with
-// with the "%#v" format string
-func (r proxyError) GoString() string {
-	return r.DebugReport()
-}
-
-// proxyError wraps another error
-type proxyError struct {
-	*TraceErr
-}
-
-type errorReport struct {
-	// OrigErrType specifies the error type as text
-	OrigErrType string
-	// OrigErrMessage specifies the original error's message
-	OrigErrMessage string
-	// Fields lists any additional fields attached to the error
-	Fields map[string]interface{}
-	// StackTrace specifies the call stack
-	StackTrace string
-	// UserMessage is the user-facing message (if any)
-	UserMessage string
-	// Caught optionally specifies the stack trace where the error
-	// has been recorded after coming over the wire
-	Caught string
-}
-
-var reportTemplate = template.Must(template.New("debugReport").Parse(reportTemplateText))
-var reportTemplateText = `
-ERROR REPORT:
-Original Error: {{.OrigErrType}} {{.OrigErrMessage}}
-{{if .Fields}}Fields:
-{{range $key, $value := .Fields}}  {{$key}}: {{$value}}
-{{end}}{{end}}Stack Trace:
-{{.StackTrace}}
-{{if .Caught}}Caught:
-{{.Caught}}
-User Message: {{.UserMessage}}
-{{else}}User Message: {{.UserMessage}}{{end}}`
